Fix error in routing to new doc in sidebar
There is a bug in Docusaurus. Only 3 letters are used for hashing the routes/paths. A hash collision can occur when adding a new doc to a group of docs in sidebars.tsx. See route.js
A workaround is adding the new file and then changing the name of said file. SSO file was added to the projects group in Sidebars.tsx.