refactor(homepagefeatures): com-357 update overview of standards with latest
Update the overview of standards on the front page (#118)

Made the overview of the standards comply with the design, showing only 7 standards. Removed the outdated standards that were not yet rewritten in accordance with the format and added the standard about Secure coding as that is an important standard. Also removed the reference in the sidebar to the outdated standard about third party dependencies.